main: Change return type to Result<(), &'static str>
Instead of panicking in situations where execution was not successful,
it's generally more productive to properly print an error and exit with
a non-zero status code.  (Panics are generally not good UX, and should
really be reserved for cases where we're fully "out of ideas".)

First, adjust the return signature of main away from returning () (the
default) towards instead returning a static, hard-coded str type.  I may
need to revisit the exact type, but this is the simplest I can come up
with.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>